[Executorch][sdpa] Refactor sdpa into impl and op
Mostly preparing for adding quantized sdpa

Differential Revision: [D71370596](https://our.internmc.facebook.com/intern/diff/D71370596/)

[ghstack-poisoned]
